Change to a ExperimentClient Broker

+++ tests/integration/service/test_service.py
@@ -3,6 +3,8 @@
 import multiprocessing
 import os
 import signal
+import time
+import logging
 from contextlib import contextmanager
 
 from orion.service.client import ClientREST
@@ -10,6 +12,15 @@ from orion.service.client import ClientREST
 TOKEN = "Tok1"
 TOKEN2 = "Tok2"
 ENDPOINT = "http://localhost:8080"
+
+log = logging.getLogger(__file__)
+
+
+def wait(p):
+    acc = 0
+    while p.is_alive() and acc < 2:
+        acc += 0.01
+        time.sleep(0.01)
 
 
 @contextmanager
@@ -27,20 +38,34 @@ def server() -> None:
     try:
         yield p
     finally:
+        # raise KeyboardInterrupt for regular shutdown
         os.kill(p.pid, signal.SIGINT)
-        p.terminate()
+        wait(p)
+
+        if p.is_alive():
+            log.debug("process still alive after sigint")
+            # notify the process we want to terminate it with SIGTERM
+            p.terminate()
+            wait(p)
+
+        if p.is_alive():
+            log.debug("process still alive after sigterm")
+            # process is taking too long kill it
+            p.kill()
+            wait(p)
+
+        p.join()
 
 
 def test_new_experiment():
     with server():
         client = ClientREST(ENDPOINT, TOKEN)
-        expid = client.new_experiment(name="MyExperiment", space=dict(a="uniform(0, 1)", b="uniform(0, 1)"))
-        print(expid)
+        expid1 = client.new_experiment(name="MyExperiment", space=dict(a="uniform(0, 1)", b="uniform(0, 1)"))
 
         # same experiment should be no problem
         client2 = ClientREST(ENDPOINT, TOKEN2)
-        expid = client.new_experiment(name="MyExperiment", space=dict(a="uniform(0, 1)", b="uniform(0, 1)"))
-        print(expid)
+        expid2 = client.new_experiment(name="MyExperiment", space=dict(a="uniform(0, 1)", b="uniform(0, 1)"))
+        assert expid1 != expid2
 
 
 def test_suggest():
